Some more test cleanup: * Move DiffUtilSpec from it to test config * Use TestKitFix instead of AkkaTestUtil * ClassfileDepickler spec uses EnsimeVFS fixture
