+++ core/src/test/scala/org/ensime/util/DiffUtilSpec.scala
