+++ server/src/test/scala/org/ensime/server/WebServerSpec.scala
@@ -22,7 +22,6 @@ import scala.xml.NodeSeq
 import spray.json._
 
 import org.ensime.api._
-import org.ensime.core._
 
 import org.ensime.util.file._
 
@@ -108,11 +107,20 @@ class WebServerSpec extends HttpFlatSpec with WebServer {
  * http://doc.akka.io/docs/akka-stream-and-http-experimental/1.0/scala/http/routing-dsl/testkit.html
  */
 abstract class HttpFlatSpec
-    extends FlatSpecLike with SetupAndTearDownSystem
+    extends FlatSpecLike with BeforeAndAfterAll
     with ScalatestRouteTest
     with TestKitBase with DefaultTimeout with ImplicitSender
     with Matchers with SLF4JLogging {
   def actorRefFactory = system
   implicit val routeTimeout: RouteTestTimeout = RouteTestTimeout(timeout.duration.dilated)
   implicit val mat = ActorMaterializer()
+
+  override protected def beforeAll(): Unit = {
+    super.beforeAll()
+  }
+  override protected def afterAll(): Unit = {
+    super.afterAll()
+    TestKit.shutdownActorSystem(system)
+  }
+
 }
