+++ server/src/test/scala/org/ensime/server/WebSocketBoilerplateSpec.scala
@@ -4,23 +4,23 @@ package org.ensime.server
 
 import akka.actor._
 import akka.http.scaladsl.model.ws._
+import akka.pattern.pipe
 import akka.stream._
 import akka.stream.scaladsl._
 import akka.testkit._
 import akka.util.ByteString
-import org.ensime.core.AkkaFlatSpec
-import org.scalatest._
+import org.ensime.fixture.SharedTestKitFixture
+import org.ensime.util.EnsimeSpec
 import spray.json._
 
-class WebSocketBoilerplateSpec extends AkkaFlatSpec {
+class WebSocketBoilerplateSpec extends EnsimeSpec with SharedTestKitFixture {
   import WebSocketBoilerplate._
 
-  implicit val mat = ActorMaterializer()
+  "WebSocketBoilerplate" should "produce Flow[In, Out]" in withTestKit { tk =>
+    import tk.system
+    import tk.system.dispatcher
+    implicit val mat = ActorMaterializer()
 
-  import system.dispatcher
-  import akka.pattern.pipe
-
-  "WebSocketBoilerplate" should "produce Flow[In, Out]" in {
     val service = TestProbe()
 
     var target: ActorRef = null
@@ -52,7 +52,11 @@ class WebSocketBoilerplateSpec extends AkkaFlatSpec {
   val foo = Foo("hello")
   val bar = Bar(13L)
 
-  it should "produce a marshalled Flow that accepts valid messages" in {
+  it should "produce a marshalled Flow that accepts valid messages" in withTestKit { tk =>
+    import tk.system
+    import tk.system.dispatcher
+    implicit val mat = ActorMaterializer()
+
     // This is quite horrible and really highlights why a BidiFlow
     // model would be better. WebSockets are *not* request / response
     // (like this).
@@ -70,7 +74,11 @@ class WebSocketBoilerplateSpec extends AkkaFlatSpec {
     client.expectMsg(TextMessage(bar.toJson.prettyPrint))
   }
 
-  it should "produce a marshalled Flow that errors on bad message" in {
+  it should "produce a marshalled Flow that errors on bad message" in withTestKit { tk =>
+    import tk.system
+    import tk.system.dispatcher
+    implicit val mat = ActorMaterializer()
+
     val user = Flow[Foo].map { f =>
       f shouldBe foo
       bar
@@ -87,7 +95,11 @@ class WebSocketBoilerplateSpec extends AkkaFlatSpec {
     }
   }
 
-  it should "produce a marshalled Flow that errors on bad inbound JSON" in {
+  it should "produce a marshalled Flow that errors on bad inbound JSON" in withTestKit { tk =>
+    import tk.system
+    import tk.system.dispatcher
+    implicit val mat = ActorMaterializer()
+
     val user = Flow[Foo].map { _ => bar }
     val endpoints = jsonMarshalledMessageFlow(user)
